# Break cyclic dependencies in `libparsec_protocol`