--- a/oxidation/libparsec/crates/client_connection/src/anonymous_cmds.rs
+++ b/oxidation/libparsec/crates/client_connection/src/anonymous_cmds.rs
@@ -1,6 +1,6 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 (eventually AGPL-3.0) 2016-present Scille SAS
 
-use libparsec_protocol::Request;
+use libparsec_miniprotocol::Request;
 use libparsec_types::BackendAnonymousAddr;
 use reqwest::{
     header::{HeaderMap, HeaderValue, CONTENT_LENGTH, CONTENT_TYPE},
@@ -54,7 +54,7 @@ impl AnonymousCmds {
     pub async fn send<T>(
         &self,
         request: T,
-    ) -> CommandResult<<T as libparsec_protocol::Request>::Response>
+    ) -> CommandResult<<T as libparsec_miniprotocol::Request>::Response>
     where
         T: Request,
     {