--- a/oxidation/libparsec/crates/protocol/tests/test_user.rs
+++ b/oxidation/libparsec/crates/protocol/tests/test_user.rs
@@ -3,7 +3,8 @@
 use hex_literal::hex;
 use std::num::NonZeroU64;
 
-use libparsec_protocol::{authenticated_cmds::v2 as authenticated_cmds, IntegerBetween1And100};
+use libparsec_miniprotocol::IntegerBetween1And100;
+use libparsec_protocol::authenticated_cmds::v2 as authenticated_cmds;
 use libparsec_tests_fixtures::parsec_test;
 use libparsec_types::HumanHandle;