--- a/oxidation/libparsec/src/lib.rs
+++ b/oxidation/libparsec/src/lib.rs
@@ -5,7 +5,10 @@ pub use libparsec_core_fs as core_fs;
 #[cfg(all(not(target_arch = "wasm32"), feature = "test-utils"))]
 pub use libparsec_platform_local_db as local_db;
 #[cfg(not(target_arch = "wasm32"))]
-pub use libparsec_protocol as protocol;
+pub mod protocol {
+    pub use libparsec_miniprotocol::{IntegerBetween1And100, Request};
+    pub use libparsec_protocol::*;
+}
 #[cfg(not(target_arch = "wasm32"))]
 pub use libparsec_types as types;