The crate `libparsec_protocol` depends on `libparsec_serialization_format` which it self depends on `libparsec_protocol::Request` (in dev-dependencies for testing purpose).

To resolve that, I've create an intermediate crate `libparsec_miniprotocol` that will be used by the crate `protocol` and `serialization_format`.

The Gain
--------

When compiling `libparsec_protocol[test]` the gain is not obvious:

```
master: 11.1s
this branch: 10.4s
```

But when compiling `libparsec_serialization_format[test]` the gain is clear:

```
master: 12.5s
this branch: 2.7s
```

This gain is caused because `libparsec_protocol` take a lot of time in `code-gen` (=> generating code) and `serialization_format` only required `protocol` for `Request`

> Those result is obtain with
>
> ```bash
> cargo clean -p <crate>
> cargo build -timings -p <crate> --tests
> ```

Others Changes
--------------

- Updated tests to use `libparsec_miniprotocol` when needed.
- Rework how we expose `libparsec::protocol`.
- Update `serialization_format` to use `libparsec_miniprotocol` when generating code.